refactor: extract image reading logic into shared module
Move detectMediaType, optimizeWithSips, and readImageFile into
assistant/src/tools/shared/filesystem/image-read.ts. Refactor
view-image.ts to delegate to the shared module. This prepares for
folding image support into file_read in a follow-up PR.
